+++ internal/provider/resource_ise_internal_user.go
@@ -200,7 +200,12 @@ func (r *InternalUserResource) Read(ctx context.Context, req resource.ReadReques
 		return
 	}
 
-	state.updateFromBody(ctx, res)
+	// If every attribute is set to null we are dealing with an import operation and therefore reading all attributes
+	if state.isNull(ctx, res) {
+		state.fromBody(ctx, res)
+	} else {
+		state.updateFromBody(ctx, res)
+	}
 
 	tflog.Debug(ctx, fmt.Sprintf("%s: Read finished successfully", state.Id.ValueString()))