+++ internal/provider/resource_ise_network_access_authorization_exception_rule.go
@@ -302,7 +302,12 @@ func (r *NetworkAccessAuthorizationExceptionRuleResource) Read(ctx context.Conte
 		return
 	}
 
-	state.updateFromBody(ctx, res)
+	// If every attribute is set to null we are dealing with an import operation and therefore reading all attributes
+	if state.isNull(ctx, res) {
+		state.fromBody(ctx, res)
+	} else {
+		state.updateFromBody(ctx, res)
+	}
 
 	tflog.Debug(ctx, fmt.Sprintf("%s: Read finished successfully", state.Id.ValueString()))