+++ internal/provider/resource_ise_trustsec_ip_to_sgt_mapping.go
@@ -173,7 +173,12 @@ func (r *TrustSecIPToSGTMappingResource) Read(ctx context.Context, req resource.
 		return
 	}
 
-	state.updateFromBody(ctx, res)
+	// If every attribute is set to null we are dealing with an import operation and therefore reading all attributes
+	if state.isNull(ctx, res) {
+		state.fromBody(ctx, res)
+	} else {
+		state.updateFromBody(ctx, res)
+	}
 
 	tflog.Debug(ctx, fmt.Sprintf("%s: Read finished successfully", state.Id.ValueString()))