+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.1.10 (unreleased)
 
 - Retry on 400 and 500 HTTP errors when creating or updating objects
+- Fix import operation of resources
 
 ## 0.1.9