+++ docs/guides/changelog.md
@@ -10,6 +10,7 @@ description: |-
 ## 0.1.10 (unreleased)
 
 - Retry on 400 and 500 HTTP errors when creating or updating objects
+- Fix import operation of resources
 
 ## 0.1.9